quant

+++ app/Http/Livewire/Sell/OrderSellDetail.php
@@ -200,7 +200,7 @@ class OrderSellDetail extends Component
         Config::set('database.connections.other.database', Auth::user()->company);
         return [
             'item' => ['required','integer','gt:0', 'exists:other.items,item_no'],
-            'Quant' =>   ['required','integer','gt:0'],
+            'Quant' =>   ['required','numeric','gt:0'],
             'price' =>   ['required','numeric'  ,'gt:0'],
         ];
     }
